Remove SKIP command
This command is to tell SA to "skip" this connection, it's for when the
"client opened connection then changed its mind".

It doesn't make sense for us to have this, since there is never a case
where we open a connection and then "change our mind" since we don't
dial the SA server until we call a command.

Maybe this will change in the future, in which case we can add it again,
but for now let's just remove it.